+++ x/programs/rust/examples/automated-market-maker/src/lib.rs
@@ -3,10 +3,9 @@
 
 use std::cmp;
 use token::Units;
-use wasmlanche_sdk::{public, state_schema, Context, ExternalCallContext, Gas, Id, Program};
+use wasmlanche_sdk::{public, state_schema, Context, ExternalCallContext, Gas, Program};
 
 mod math;
-mod tests;
 
 state_schema! {
     // Tokens in the Pool as Program type
@@ -20,7 +19,7 @@ const MAX_GAS: Gas = 10000000;
 
 /// Initializes the pool with the two tokens and the liquidity token
 #[public]
-pub fn init(context: &mut Context, token_x: Program, token_y: Program, liquidity_token: Id) {
+pub fn init(context: &mut Context, token_x: Program, token_y: Program, liquidity_token: Program) {
     context
         .store((
             (TokenX, token_x),
@@ -31,8 +30,10 @@ pub fn init(context: &mut Context, token_x: Program, token_y: Program, liquidity
 
     let liquidity_context = ExternalCallContext::new(liquidity_token, MAX_GAS, 0);
 
-    // initialize the liquidity token
-    // let init_result = context
+    let transfer_reuslt =
+        token::transfer_ownership(&liquidity_context, *context.program().account());
+    // TODO: the init function should spin up a new token contract instead
+    // of requiring the caller to pass in the liquidity token
     assert!(
         transfer_reuslt,
         "failed to transfer ownership of the liquidity token"
