+++ x/programs/rust/wasmlanche-sdk/src/program.rs
@@ -1,7 +1,7 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-use crate::{memory::HostPtr, types::Address, types::Id, Gas};
+use crate::{memory::HostPtr, types::Address, Gas};
 use borsh::{BorshDeserialize, BorshSerialize};
 use thiserror::Error;
 
@@ -117,24 +117,6 @@ impl Program {
         let bytes = unsafe { get_remaining_fuel() };
 
         borsh::from_slice::<u64>(&bytes).expect("failed to deserialize the remaining fuel")
-    }
-
-    /// Deploy an instance of the specified program and returns the account of the new instance
-    /// # Panics
-    /// Panics if there was an issue deserializing the account
-    #[must_use]
-    pub fn deploy(&self, program_id: Id, account_creation_data: &[u8]) -> Address {
-        #[link(wasm_import_module = "program")]
-        extern "C" {
-            #[link_name = "deploy"]
-            fn deploy(ptr: *const u8, len: usize) -> HostPtr;
-        }
-        let ptr =
-            borsh::to_vec(&(program_id, account_creation_data)).expect("failed to serialize args");
-
-        let bytes = unsafe { deploy(ptr.as_ptr(), ptr.len()) };
-
-        borsh::from_slice(&bytes).expect("failed to deserialize the account")
     }
 }
