+++ x/programs/test/programs/deploy_program/src/lib.rs
@@ -5,5 +5,5 @@ use wasmlanche_sdk::{public, Address, Context, Id};
 
 #[public]
 pub fn deploy(ctx: &mut Context, program_id: Id) -> Address {
-    ctx.program().deploy(program_id, &[])
+    ctx.deploy(program_id, &[])
 }
